LPS-62022 Rename package and add it to portal-test/bnd.bnd.
