+++ portal-test/bnd.bnd
@@ -2,6 +2,7 @@ Bundle-Name: Liferay Portal Test
 Bundle-Version: 7.0.0
 Export-Package:\
 	com.liferay.portal.kernel.test,\
+	com.liferay.portal.kernel.test.context,\
 	com.liferay.portal.kernel.test.junit,\
 	com.liferay.portal.kernel.test.randomizerbumpers,\
 	com.liferay.portal.kernel.test.rule,\
