+++ portal-test/src/com/liferay/portal/kernel/test/context/ContextUserReplace.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.test;
+package com.liferay.portal.kernel.test.context;
 
 import com.liferay.portal.model.User;
 import com.liferay.portal.security.auth.PrincipalThreadLocal;
